+++ wayland.h
@@ -41,6 +41,37 @@ struct monitor {
     float inch;  /* e.g. 24" */
 };
 
+enum binding_action {
+    BIND_ACTION_NONE,
+    BIND_ACTION_SCROLLBACK_UP,
+    BIND_ACTION_SCROLLBACK_DOWN,
+    BIND_ACTION_CLIPBOARD_COPY,
+    BIND_ACTION_CLIPBOARD_PASTE,
+    BIND_ACTION_PRIMARY_PASTE,
+    BIND_ACTION_SEARCH_START,
+    BIND_ACTION_FONT_SIZE_UP,
+    BIND_ACTION_FONT_SIZE_DOWN,
+    BIND_ACTION_FONT_SIZE_RESET,
+    BIND_ACTION_SPAWN_TERMINAL,
+    BIND_ACTION_MINIMIZE,
+    BIND_ACTION_MAXIMIZE,
+    BIND_ACTION_FULLSCREEN,
+    BIND_ACTION_COUNT,
+};
+
+struct key_binding {
+    xkb_mod_mask_t mods;
+    xkb_keysym_t sym;
+    enum binding_action action;
+};
+typedef tll(struct key_binding) key_binding_list_t;
+
+struct mouse_binding {
+    uint32_t button;
+    int count;
+    enum binding_action action;
+};
+
 struct kbd {
     struct xkb_context *xkb;
     struct xkb_keymap *xkb_keymap;
@@ -66,6 +97,11 @@ struct kbd {
     bool alt;
     bool ctrl;
     bool meta;
+
+    struct {
+        key_binding_list_t key;
+        key_binding_list_t search;
+    } bindings;
 };
 
 struct wl_clipboard {